Move `Component.onCompleted` at the end
This is used to trigger an action when the page has finished being built
and is now ready for use.  Instead of having this code at the end of the
definition of the SilicaListView at the end of the SilicaFlickable, put
it directly at the end of the Page itself.

In next commits, we are going to move code around, and this will make
maintainers life easier when reviewing changes by minimizing the size of
the diffs.