Move `PullDownMenu` into the `SilicaListView`
A `SilicaListView` will properly handle a child `PullDownMenu`.  It was
previously working when it was the child of a `SilicaFlickable`, but
since it was replaced by a `Column` it was not usable anymore.

Attaching the `PullDownMenu` to the `SilicaListView` instead of the
`SilicaFlickable`, it is properly hidden when the list of OTP is scroll
down and the menu cannot be operated by scrolling down.